wip: odmx: add function to check for existing equipment and data to equipment maps because we should do this for all data source types not just decagon; move check_diff to write_equipment_jsons, remove no-longer-used get_mapping function
